Add support for generating the pots themselves.
Previously this scad file could only generate the rack or drip tray.
Now it can generate a pot() that goes into the rack too.

There's a big hole in the bottom of the pots still, so I need to
add a bit of code to generate the pot's bottom inserts too next.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>